+++ pkgbuilder/exceptions.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python3
 # -*- encoding: utf-8 -*-
-# PKGBUILDer v4.0.3
+# PKGBUILDer v4.1.0
 # An AUR helper (and library) in Python 3.
 # Copyright © 2011-2015, Chris Warrick.
 # See /LICENSE for licensing information.
@@ -225,6 +225,7 @@ class CloneError(PBException):
         self.exit = exit
         self.args = args
         self.kwargs = kwargs
+
 
 class EmptyRepoError(CloneError):